added better styles

+++ delete.php
@@ -2,7 +2,7 @@
   include 'connection.php';
 
 
-  if (isset($_GET['deleteid'])) {
+  if (isset($_GET['deleteid'])&&!(isset($_GET['table']))) {
     $id=$_GET['deleteid'];
 
     $sql="delete from student where id=$id";
@@ -16,7 +16,6 @@
   if (isset($_GET['deleteid'])&&isset($_GET['table'])) {
     $id=$_GET['deleteid'];
     $table=$_GET['table'];
-
     $sql="delete from $table where id=$id";
     $result=mysqli_query($con,$sql);
     if($result){
